refactor: code and styles
