--- .travis.yml
@@ -1,4 +1,0 @@
-language: node_js
-node_js:
-  - 9
-  - 8
