+++ example/src/App.js
@@ -10,7 +10,7 @@ export default class App extends Component {
           app='good-books-ds'
           credentials='nY6NNTZZ6:27b76b9f-18ea-456c-bc5e-3a5263ebc63d'
           dataField={['original_title', 'original_title.search']}
-          themePreset='dark'
+          showVoiceSearch
         />
       </div>
     );
