+++ rollup.config.js
@@ -1,12 +1,12 @@
-import babel from 'rollup-plugin-babel'
-import commonjs from 'rollup-plugin-commonjs'
-import external from 'rollup-plugin-peer-deps-external'
-import postcss from 'rollup-plugin-postcss'
-import resolve from 'rollup-plugin-node-resolve'
-import url from 'rollup-plugin-url'
-import svgr from '@svgr/rollup'
+import babel from 'rollup-plugin-babel';
+import commonjs from 'rollup-plugin-commonjs';
+import external from 'rollup-plugin-peer-deps-external';
+import postcss from 'rollup-plugin-postcss';
+import resolve from 'rollup-plugin-node-resolve';
+import url from 'rollup-plugin-url';
+import svgr from '@svgr/rollup';
 
-import pkg from './package.json'
+import pkg from './package.json';
 
 export default {
   input: 'src/index.js',
@@ -15,8 +15,7 @@ export default {
     '@emotion/core',
     '@emotion/styled',
     'react-highlight-words',
-    '@appbaseio/searchbase',
-    'cross-fetch'
+    '@appbaseio/searchbase'
   ],
   output: [
     {
@@ -43,4 +42,4 @@ export default {
     resolve(),
     commonjs()
   ]
-}
+};
