+++ src/styles/Input.js
@@ -1,13 +1,9 @@
 import { css } from '@emotion/core';
 import styled from '@emotion/styled';
+import { dark } from './Theme';
 
 const alertBorder = ({ theme }) => css`
   border: 1px solid ${theme.colors.alertColor};
-`;
-
-const webkitLineClamp = css`
-  line-height: 1.4em;
-  max-height: 5.6em;
 `;
 
 const input = css`
@@ -23,10 +19,6 @@ const input = css`
   &:focus {
     background-color: #fff;
   }
-`;
-
-const dark = theme => css`
-  border-color: ${theme.colors.borderColor};
 `;
 
 const darkInput = ({ theme }) => css`
@@ -75,94 +67,5 @@ const Input = styled.input`
 
   ${props => props.alert && alertBorder};
 `;
-const noSuggestions = (themePreset, theme) => css`
-  display: block;
-  width: 100%;
-  border: 1px solid #ccc;
-  background-color: #fff;
-  font-size: 0.9rem;
-  z-index: 3;
-  position: absolute;
-  top: 39px;
-  margin: 0;
-  padding: 0;
-  list-style: none;
-  max-height: 260px;
-  overflow-y: auto;
-
-  &.small {
-    top: 30px;
-  }
-
-  li {
-    display: flex;
-    justify-content: space-between;
-    padding: 10px;
-    user-select: none;
-
-    & > .trim {
-      display: block;
-      display: -webkit-box;
-      width: 100%;
-      max-height: 2.3rem;
-      line-height: 1.2rem;
-      -webkit-box-orient: vertical;
-      overflow: hidden;
-      text-overflow: ellipsis;
-      ${webkitLineClamp}
-    }
-  }
-
-  ${themePreset === 'dark' && dark(theme)}
-`;
-
-const suggestions = (themePreset, theme) => css`
-  display: block;
-  width: 100%;
-  border: 1px solid #ccc;
-  background-color: #fff;
-  font-size: 0.9rem;
-  z-index: 3;
-  position: absolute;
-  top: 39px;
-  margin: 0;
-  padding: 0;
-  list-style: none;
-  max-height: 260px;
-  overflow-y: auto;
-
-  &.small {
-    top: 30px;
-  }
-
-  li {
-    display: flex;
-    justify-content: space-between;
-    cursor: pointer;
-    padding: 10px;
-    user-select: none;
-
-    .trim {
-      overflow: hidden;
-      text-overflow: ellipsis;
-      white-space: nowrap;
-    }
-
-    &:hover,
-    &:focus {
-      background-color: #eee;
-    }
-  }
-
-  ${themePreset === 'dark' && dark(theme)}
-`;
-
-const suggestionsContainer = css`
-  position: relative;
-  .cancel-icon {
-    cursor: pointer;
-  }
-`;
 
 export default Input;
-export { suggestionsContainer, suggestions, input, noSuggestions };
