+++ src/styles/Theme.js
@@ -1,0 +1,5 @@
+import { css } from '@emotion/core';
+
+export const dark = theme => css`
+  border-color: ${theme.colors.borderColor};
+`;
